class hierarhy was changed to relieve duplication
